+++ app/src/main/java/team/retum/jobisandroidv2/navigation/AuthNavigation.kt
@@ -3,6 +3,7 @@ package team.retum.jobisandroidv2.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
+import team.retum.landing.NAVIGATION_LANDING
 import team.retum.landing.landing
 import team.retum.signin.navigateToSignIn
 import team.retum.signin.signIn
@@ -15,7 +16,7 @@ const val NAVIGATION_AUTH = "auth"
 internal fun NavGraphBuilder.authNavigation(navController: NavController) {
     navigation(
         route = NAVIGATION_AUTH,
-        startDestination = NAVIGATION_SIGN_UP,
+        startDestination = NAVIGATION_LANDING,
     ) {
         landing(
             onSignInClick = navController::navigateToSignIn,
